% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_datahelpers.R
\name{add_cds_factor_columns}
\alias{add_cds_factor_columns}
\title{Add predefined sample-level cell metadata to cell data sets while importing}
\usage{
add_cds_factor_columns(cds, columns_to_add)
}
\arguments{
\item{cds}{A cell data set object}

\item{columns_to_add}{A named vector where the name of each element becomes the name of the new colData column and the value is the value for that particular sample.  Best used when importing from a metadata table.}
}
\description{
Add predefined sample-level cell metadata to cell data sets while importing
}
