% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-sig-table-npc-v2.R
\name{add_sig_annotations}
\alias{add_sig_annotations}
\title{Add significance annotations to an existing ggplot}
\usage{
add_sig_annotations(plot, ...)
}
\description{
Add significance annotations to an existing ggplot
}
