% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ape_class.R
\docType{class}
\name{Ape-class}
\alias{Ape-class}
\alias{Ape}
\title{An S4 class to hold genebank/APE file data.}
\description{
An instance of this class is best created by calling "bb_parseape()" on a genebank or APE-formatted file.  That function will parse the file, correctly format the sections and place them in the slots of the Ape Object.  Technically only "LOCUS" is a required slot for the Ape object, however there is no point without having "ORIGIN" (sequence data), and so bb_parseape() will fail without an "ORIGIN" section.  Other slots are optional.  Additional slots will be ignored by the constructor function.  DNA sequence will be stored in a DNAStringSet object and features in a GRanges object.  See https://www.ncbi.nlm.nih.gov/Sitemap/samplerecord.html for genebank file specification.
}
\section{Slots}{

\describe{
\item{\code{LOCUS}}{The LOCUS line of the genebank formatted as a character string.}

\item{\code{DEFINITION}}{The DEFINITION line of the genebank file formatted as a character string.}

\item{\code{ACCESSION}}{The ACCESSION section of the genebank file formatted as a character string.}

\item{\code{VERSION}}{The VERSION section of the genebank file formatted as a character string.}

\item{\code{SOURCE}}{The SOURCE section of the genebank file formatted as a character string.}

\item{\code{COMMENT}}{The COMMENT section of the genebank file formatted as a character string.}

\item{\code{FEATURES}}{The FEATURES section of the genebank file formatted as a character string.  Created internally from the GRanges object.  Caution:  some FEATURE attributes may be lost in conversion.}

\item{\code{ORIGIN}}{The DNA sequence}

\item{\code{end_of_file}}{The end of the file signal.}

\item{\code{dna_biostring}}{The entire ORIGIN sequence formatted as a DNAStringSet of length 1.}

\item{\code{granges}}{Genebank features formatted as a GRanges object.}
}}

