% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ape_class.R
\name{Ape.fasta}
\alias{Ape.fasta}
\title{Save an Ape Instance as a Fasta File}
\usage{
Ape.fasta(ape, feature = NULL, out)
}
\arguments{
\item{feature}{Name of feature to select when writing FASTA file.  If null (default), the whole biostring will be saved as a fasta.}

\item{out}{Name of FASTA file to write}
}
\description{
Save an Ape Instance as a Fasta File
}
