% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ape_class.R
\name{Ape.fimo}
\alias{Ape.fimo}
\title{Run FIMO on Selected Ape Object Features}
\usage{
Ape.fimo(ape, fimo_feature, out = NULL)
}
\arguments{
\item{ape}{An Ape instance}

\item{fimo_feature}{A character vector of features from the Ape object that will be used to run fimo.}

\item{out}{Directory that will be created to hold the fimo results.  A date/time stamp will be appended.  If null, the objects will not be saved and the function will only return a GRanges object}
}
\description{
For the supplied Ape object, run FIMO to identify putative transcription factor binding sites in a DNA subsequence.
}
