% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ape_class.R
\name{Ape.save}
\alias{Ape.save}
\title{Save an Ape Instance as a Genebank Format File}
\usage{
Ape.save(x, out)
}
\arguments{
\item{out}{Name of genebank/APE file to write}
}
\description{
Save an Ape Instance as a Genebank Format File
}
