% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_gene_umap.R
\name{bb_aggregate}
\alias{bb_aggregate}
\title{Aggregate Single Cell Gene Expression}
\usage{
bb_aggregate(
  obj,
  assay = "RNA",
  gene_group_df = NULL,
  cell_group_df = NULL,
  norm_method = c("log", "binary", "size_only"),
  pseudocount = 1,
  scale_agg_values = TRUE,
  max_agg_value = 3,
  min_agg_value = -3,
  exclude.na = TRUE
)
}
\arguments{
\item{obj}{A Seurat or cell data set object}

\item{assay}{Gene expression assay to use for aggregation; currently only applies to Seurat objects, Default: 'RNA'}

\item{gene_group_df}{A 2-column dataframe with gene names or ids and gene groupings, Default: NULL}

\item{cell_group_df}{A 2-coumn dataframe with cell ids and gene groupings, Default: NULL}

\item{norm_method}{Gene normalization method, Default: c("log", "binary", "size_only")}

\item{pseudocount}{Pseudocount, Default: 1}

\item{scale_agg_values}{Whether to scale the aggregated values, Default: TRUE}

\item{max_agg_value}{If scaling, make this the maximum aggregated value, Default: 3}

\item{min_agg_value}{If scaling, make this the minimum aggregated value, Default: -3}

\item{exclude.na}{Exclude NA?, Default: TRUE}
}
\value{
A dense or sparse matrix.
}
\description{
Generates a matrix of counts aggregated by gene and/or cell group.
}
\details{
The best way to group genes or cells is by using bb_*meta and then select cell_id or feature_id plus one metadata column with your group labels.
}
\seealso{
\code{\link[cli]{cli_div}}, \code{\link[cli]{cli_alert}}
\code{\link[monocle3]{normalized_counts}}, \code{\link[monocle3]{my.aggregate.Matrix}}
\code{\link[Matrix]{character(0)}}
}
