% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_align.R
\name{bb_align}
\alias{bb_align}
\title{Align a CDS object according to a metadata variable}
\usage{
bb_align(cds, align_by, n_cores = 8)
}
\arguments{
\item{cds}{A cell data set object to align}

\item{align_by}{A metadata column to align by}

\item{n_cores}{Number of cores to reduce dimensions by.}
}
\value{
A modified cell data set object with aligned dimensions and new metadata columns holding prealignment umap coordinates.
}
\description{
Align a CDS object according to a metadata variable
}
