% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_annotate_npc.R
\name{bb_annotate_npc}
\alias{bb_annotate_npc}
\title{Annotate a Plot using NPC Coordinates}
\usage{
bb_annotate_npc(label, x, y, ...)
}
\arguments{
\item{label}{the text label to apply to the plot}

\item{x}{NPC X coordinate}

\item{y}{NPC Y coordinate}
}
\description{
Annotate a Plot using NPC Coordinates
}
