% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_funcs.R
\name{bb_buff_granges}
\alias{bb_buff_granges}
\title{Clean Up Your GRanges Object}
\usage{
bb_buff_granges(x, gen)
}
\arguments{
\item{x}{A Granges object to buff.}

\item{gen}{An optional genome name to provide.  Recommend "hg38" or "danRer11".}
}
\value{
A GRanges object
}
\description{
This function does several things.  It removes ranges with non-standard chromosomes and drops their levels.  It will optionally set the genome to the user-provided value.  Typically we would use "hg38" or "danRer11". This is the exported version because it is so useful.
}
