% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat.R
\name{bb_cellchat_heatmap}
\alias{bb_cellchat_heatmap}
\title{Make a Complex Heatmap From a CellChat Object With Sensible Defaults}
\usage{
bb_cellchat_heatmap(
  object,
  source_filter = NULL,
  target_filter = NULL,
  interaction_threshold = 0,
  colors = c("transparent", "red3"),
  rowanno = c(NULL, "Annotation", "Pathway"),
  rowanno_colors = NULL,
  colanno = c(NULL, "Source", "Target"),
  colanno_colors = NULL,
  pval_filter = 0.05,
  heatmap_name = "Interaction\\nScore",
  heatmap_show_row_dend = TRUE,
  heatmap_row_dend_width = unit(5, "mm"),
  heatmap_show_column_dend = TRUE,
  heatmap_column_dend_height = unit(5, "mm"),
  heatmap_row_names_gp = gpar(fontsize = 10),
  heatmap_column_names_gp = gpar(fontsize = 10),
  heatmap_column_names_rot = 90,
  heatmap_column_title = NULL,
  heatmap_column_title_gp = gpar(fontsize = 12, fontface = "bold"),
  col_anno_name_gp = gpar(fontsize = 10, fontface = "bold"),
  row_anno_name_gp = gpar(fontsize = 10, fontface = "bold")
)
}
\arguments{
\item{object}{The CellChat object to plot}

\item{source_filter}{Optional filter for source cell clusters from the object metadata.  Accepts a single string or vector of cell groups., Default: NULL}

\item{target_filter}{Optional filter for target cell clusters from the object metadata.  Accepts a single string or vector of cell groups., Default: NULL}

\item{colors}{Color scale endpoints, Default: c("transparent", "red3")}

\item{rowanno}{Options for simple row annotation; must be one of c(NULL, "Annotation", "Pathway")}

\item{rowanno_colors}{Optional colors to replace the poor color selections from Complex heatmap.  Must be supplied as a named list with one element each for "Annotation" and "Pathway".  Not required if not showing these annotations.  The list should be of the form:  list(Annotation = c("name1" = "color value1", "name2" = "color_value2")), Default: NULL}

\item{colanno}{Options for simple column annotation; must be one of c(NULL, "Source", "Target")}

\item{colanno_colors}{See rowanno_colors, Default: NULL}

\item{pval_filter}{Filter for significance of associations.  CellChat returns pvalues of 0, 0.01, and 0.05; this function will filter and retain values less than or equal to the provided value. Default: 0.05}

\item{heatmap_name}{Name for the main color scale of the heatmap, Default: 'InteractionScore'}

\item{heatmap_show_row_dend}{Show row dendrograms? Default: TRUE}

\item{heatmap_row_dend_width}{Width of row dendrograms Default: unit(5, "mm")}

\item{heatmap_show_column_dend}{Show column dendrograms?' Default: TRUE}

\item{heatmap_column_dend_height}{Height of column dendrograms. Default: unit(5, "mm")}

\item{heatmap_row_names_gp}{Row name graphical params, Default: gpar(fontsize = 10)}

\item{heatmap_column_names_gp}{Column name graphical params, Default: gpar(fontsize = 10)}

\item{heatmap_column_names_rot}{Column name rotation, Default: 90}

\item{heatmap_column_title}{Column title, Default: NULL}

\item{heatmap_column_title_gp}{Column title graphical params, Default: gpar(fontsize = 12, fontface = "bold")}

\item{col_anno_name_gp}{Column annotation name graphical params, Default: gpar(fonmtsize = 10, fontface = "bold")}

\item{row_anno_name_gp}{Row annotation name graphical params, Default: gpar(fontsize = 10, fontface = "bold")}
}
\value{
a heatmap as a grid object; plot using cowplot::plot_grid
}
\description{
This will generate heatmap from a CellChat object using ComplexHeatmap::Heatmap.  Options are provided to filter for sender and receiver cells, to generate simple marginal annotations and for aesthetic control.
}
\details{
see github::sqjin/CellChat
}
\seealso{
\code{\link[CellChat]{subsetCommunication}}
\code{\link[tibble]{as_tibble}},\code{\link[tibble]{c("tibble", "tibble")}},\code{\link[tibble]{rownames}}
\code{\link[dplyr]{filter}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}},\code{\link[dplyr]{mutate-joins}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise}}
\code{\link[tidyr]{pivot_wider}}
\code{\link[ComplexHeatmap]{rowAnnotation}},\code{\link[ComplexHeatmap]{columnAnnotation}},\code{\link[ComplexHeatmap]{draw-dispatch}},\code{\link[ComplexHeatmap]{Heatmap}}
\code{\link[circlize]{colorRamp2}}
\code{\link[grid]{grid.grab}}
}
