% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_meta.R
\name{bb_cellmeta}
\alias{bb_cellmeta}
\title{A Function to Return CDS Cell Metadata in Tibble Form}
\usage{
bb_cellmeta(cds, row_name = "cell_id")
}
\arguments{
\item{cds}{A cell data set object}

\item{row_name}{Optional name to provide for cell unique identifier.  Defaults to "cell_id"}
}
\value{
A tibble
}
\description{
A Function to Return CDS Cell Metadata in Tibble Form
}
