% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citeseq.R
\name{bb_cite_umap}
\alias{bb_cite_umap}
\title{Plot a UMAP Showing Cite-Seq Antibody Binding}
\usage{
bb_cite_umap(
  cds,
  antibody,
  assay = "CLR_counts",
  cell_size = 1,
  alpha = 1,
  alt_dim_x = NULL,
  alt_dim_y = NULL,
  plot_title = NULL,
  color_legend_title = NULL,
  order = TRUE,
  rescale = NULL,
  ncol = NULL
)
}
\arguments{
\item{cds}{The cds with an "Antibody Capture" alt experiment to plot.}

\item{antibody}{The name of the antibody to plot. Equivalent to gene_short_name.  Accepts a character vector.}

\item{assay}{The binding assay to use, Default: "CLR_counts"}

\item{cell_size}{Size of points to plot, Default: 1}

\item{alpha}{Alpha for the plotted points, Default: 1}

\item{alt_dim_x}{Alternate/reference dimensions to plot by.}

\item{alt_dim_y}{Alternate/reference dimensions to plot by.}

\item{plot_title}{Optional title for the plot, Default: NULL}

\item{color_legend_title}{Optional title for the color scale., Default: NULL}

\item{order}{Whether or not to order cells by gene expression.  When ordered, non-expressing cells are plotted first, i.e. on the bottom. Default: TRUE}

\item{rescale}{Optional redefinition of the color scale, Default: NULL}

\item{ncol}{If specified, the number of columns for facet_wrap, Default: NULL}
}
\value{
a ggplot
}
\description{
Requires a cds with an alt experiment established.  Use bb_split_citeseq to generate this and to normalize binding data using the CLR method.  Returns a ggplot.
}
\seealso{
\code{\link[SingleCellExperiment]{reducedDims}}
}
