% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_fix_file_path.R
\name{bb_fix_file_path}
\alias{bb_fix_file_path}
\title{Convert windows filepath to linux-compatible}
\usage{
bb_fix_file_path(x)
}
\arguments{
\item{x}{A character string filepath copied from Windows}
}
\value{
A linux-compatible filepath
}
\description{
Convert windows filepath to linux-compatible
}
