% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_gene_dotplot.R
\name{bb_gene_dotplot}
\alias{bb_gene_dotplot}
\title{Make a dotplot of gene expression by cell population}
\usage{
bb_gene_dotplot(
  cds,
  markers,
  group_cells_by,
  reduction_method = "UMAP",
  norm_method = c("log", "size_only"),
  lower_threshold = 0,
  max.size = 10,
  group_ordering = "bicluster",
  gene_ordering = NULL,
  pseudocount = 1,
  scale_max = 3,
  scale_min = -3,
  colorscale_name = NULL,
  sizescale_name = NULL,
  ...
)
}
\arguments{
\item{cds}{A cell data set object}

\item{markers}{A character vector of genes to plot}

\item{group_cells_by}{A cds colData column.  Use "multifactorial" to pick 2 categorical variables to put on X axis and to facet by.  See ordering below.}

\item{norm_method}{How to normalize gene expression.  Default = "log"}

\item{lower_threshold}{Lower cutoff for gene expression}

\item{max.size}{The maximum size of the dotplot}

\item{group_ordering}{Defaults to "biclustering" method from pheatmap.  Optionally will take a vector of group values to set the axis order explicitly.  If using group_cells_by = "multifactorial" you will need a df to define facet and axis levels.  See example.}

\item{pseudocount}{Add to zero expressors.  Default = 1}

\item{scale_max}{Expression scale max}

\item{scale_min}{Expression scale min}

\item{colorscale_name}{Label for the color scale}

\item{sizescale_name}{Label for the size scale}

\item{...}{Additional parameters to pass to facet_wrap.}
}
\value{
A ggplot
}
\description{
Make a dotplot of gene expression by cell population
}
\examples{
bb_gene_dotplot(
  cds_anno_aligned_tissue_id,
  group_cells_by = "multifactorial",
  markers = c("FLT3", "CXCL8", "IL3RA", "CSF2RA", "IL6R"),
  ordering = tribble(
    ~aesthetic, ~variable, ~value, ~level,
    "facet","chemo_response","sensitive",1,
    "facet","chemo_response","unresponsive",2,
    "axis","timepoint_binary","pre-treatment",1,
    "axis","timepoint_binary","post-treatment",2)
  )
}
