% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_gene_modules.R
\name{bb_gene_modules}
\alias{bb_gene_modules}
\title{A function to generate gene modules and add them to the CDS rowData}
\usage{
bb_gene_modules(cds, n_cores = 8)
}
\arguments{
\item{cds}{A cell data set object}

\item{n_cores}{Number of processor cores to use for the analysis}
}
\value{
A modified cell data set object with a new rowData column holding gene module data.
}
\description{
Based on Monocle3's gene module functions.  Implemented with default values.
}
