% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_gene_pseudotime.R
\name{bb_gene_pseudotime}
\alias{bb_gene_pseudotime}
\title{Plot expression of a gene or genes in pseudotime.}
\usage{
bb_gene_pseudotime(
  cds_subset,
  min_expr = NULL,
  cell_size = 0.75,
  nrow = NULL,
  ncol = 1,
  panel_order = NULL,
  color_cells_by = "pseudotime",
  trend_formula = "~ splines::ns(pseudotime, df=3)",
  label_by_short_name = TRUE,
  vertical_jitter = NULL,
  horizontal_jitter = NULL
)
}
\arguments{
\item{cds_subset}{A cell data set object subset with only cells and genes of interest}

\item{min_expr}{Lower threshold of expression for plotting}

\item{cell_size}{Size of point for plotting}

\item{nrow}{Number of rows for facetting}

\item{ncol}{Number of columns for facetting}

\item{panel_order}{Character string for order of genes to plot}

\item{color_cells_by}{A cds colData column}

\item{trend_formula}{Formula for the trend line}

\item{label_by_short_name}{Boolean to label by gene name or ID}

\item{vertical_jitter}{Adjustment to vertical jitter.  Optional}

\item{horizontal_jitter}{Adjustment to horizontal jitter.  Optional}
}
\value{
A ggplot
}
\description{
Plot expression of a gene or genes in pseudotime.
}
