% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_gene_umap.R
\name{bb_gene_umap}
\alias{bb_gene_umap}
\title{Make a plot of gene expression in UMAP form}
\usage{
bb_gene_umap(
  cds,
  gene_or_genes,
  cell_size = 1,
  alpha = 1,
  ncol = NULL,
  plot_title = NULL,
  color_legend_title = NULL
)
}
\arguments{
\item{cds}{A cell data set object}

\item{gene_or_genes}{A character vector of genes to plot}

\item{cell_size}{Size of cells to plot}

\item{alpha}{Alpha value for overlaid points}

\item{ncol}{Number of columns in facetted plot}

\item{plot_title}{Main title of the plot}

\item{color_legend_title}{Title for the color legend}
}
\value{
A ggplot
}
\description{
Make a plot of gene expression in UMAP form
}
