% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_gene_violinplot.R
\name{bb_gene_violinplot}
\alias{bb_gene_violinplot}
\title{Make a plot of gene expression in UMAP form}
\usage{
bb_gene_violinplot(
  cds,
  variable,
  genes_to_plot,
  pseudocount = 1,
  include_jitter = FALSE,
  ytitle = "Expression",
  plot_title = NULL,
  rows = 1,
  show_x_label = TRUE,
  legend_pos = "none",
  comparison_list = NULL,
  palette = NULL,
  violin_alpha = 1,
  jitter_alpha = 1,
  jitter_color = "black",
  jitter_fill = "transparent",
  jitter_size = 0.5,
  facet_scales = "fixed",
  order_genes = TRUE,
  jitter_match = F
)
}
\arguments{
\item{cds}{A cell data set object}

\item{variable}{Stratification variable for x-axis}

\item{genes_to_plot}{Either a character vector of gene short names or a tbl/df where the first column is gene short name and the second is the gene grouping.}

\item{pseudocount}{Value to add to zero-cells}

\item{include_jitter}{Include jitter points}

\item{ytitle}{Title for y axis}

\item{plot_title}{Main title for the plot}

\item{rows}{Number of rows for facetting}

\item{show_x_label}{Option to show x label}

\item{legend_pos}{Position for label}

\item{comparison_list}{Optional list of comparisons for ggpubr}

\item{palette}{Color palette to use.  Viridis is default.}

\item{violin_alpha}{Alpha value for violin plot}

\item{jitter_alpha}{Alpha value for jitter plot}

\item{jitter_color}{Color for the jitter plot.  Defaults to black and ignored if jitter_match == TRUE}

\item{jitter_fill}{Fill for the jitter plot}

\item{jitter_size}{Size of the jitter points}

\item{facet_scales}{Scale option for facetting.  "Fixed" is default}

\item{order_genes}{If true, put genes in the same order as variable parameter}

\item{jitter_match}{If true, match jitter color to violin fill.}
}
\value{
A ggplot
}
\description{
Make a plot of gene expression in UMAP form
}
