% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_genebubbles.R
\name{bb_genebubbles}
\alias{bb_genebubbles}
\title{Create a Gene Bubble/Dot Plot}
\usage{
bb_genebubbles(
  obj,
  genes,
  cell_grouping,
  scale_expr = TRUE,
  gene_ordering = c("bicluster", "as_supplied"),
  group_ordering = c("bicluster", "as_supplied"),
  return_value = c("plot", "data")
)
}
\arguments{
\item{obj}{A Seurat or cell_data_set object.}

\item{genes}{Gene or genes to plot.}

\item{cell_grouping}{Cell metadata column to group cells by.  Supply more than one in a vector to generate a composite variable.}

\item{scale_expr}{Whether to scale expression by gene, Default: TRUE}

\item{gene_ordering}{By default, genes will be ordered by a clustering algorithm.  Supply "as_supplied" to plot the genes in the order supplied to the "genes" argument , Default: c("bicluster", "as_supplied")}

\item{group_ordering}{By default, cell groups will be ordered by a clustering algorithm.  Supply "as_supplied" to plot the cell groups in the order supplied to "cell_grouping", Default: c("bicluster", "as_supplied")}

\item{return_value}{Whether to return a plot or data in tibble form, Default: c("plot", "data")}
}
\value{
A ggplot or a tibble
}
\description{
This is a very data-dense plot and is the recommended way for showing expression of single markers/genes by cell group.  By default, this function will return an unfaceted ggplot with cell groups on the X axis and genes on the Y axis with dot size representing proportion of cells in the cell group expressing a gene and color scale representing per-cell expression.

But it also may be of interest to add aesthetic variables such as facets or additional color scales.  There are two ways this function will facilitate that.  First, you can supply a vector of cell groups to the cell_grouping argument and a the cells will be grouped by the composite value of these factors.  Usually if you are doing this, you also will want to have access to the components of this composite variable to facet by.  So you can supply "data" to the return_value argument to get a tibble.  From there you can modify as necessary and generate a ggplot assigning aesthetics and scales as desired and using geom_point.
}
