% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_goscatter.R
\name{bb_goscatter}
\alias{bb_goscatter}
\title{Make a scatter plot of GO term associations}
\usage{
bb_goscatter(
  simMatrix,
  reducedTerms,
  size = "score",
  addLabel = TRUE,
  labelSize = 4
)
}
\arguments{
\item{simMatrix}{}

\item{labelSize}{}
}
\value{
A ggplot
}
\description{
Make a scatter plot of GO term associations
}
