% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_funcs.R
\name{bb_import_bw}
\alias{bb_import_bw}
\title{Import a Big Wig File}
\usage{
bb_import_bw(path, group, coverage_column = "score")
}
\arguments{
\item{path}{file path to the bigwig file}

\item{group}{a label to apply that describes the source data for the track}

\item{coverage_column}{PARAM_DESCRIPTION, Default: 'score'}
}
\value{
a granges object
}
\description{
This is a thin wrapper around rtracklayer::import.bw.  The purpose is to serve as a helper function for making Trace objects from bigwig files.  This function converts the name of the numeric value metadata column to "coverage" for consistency with Trace objects made from Seurat/Signac objects.  The numeric column it will look for is "score" by default.  This appears to be the default name applied by import.bw and so this is the default value for this function.  The option is provided to change it if necessary.  The group variable must be supplied.  Typically this will be something informative, like "ATAC" or some Histone mark that the data come from.
}
