% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_load_cloud_counts.R
\name{bb_load_cloud_counts}
\alias{bb_load_cloud_counts}
\title{Load 10X Cloud Pipestances Into CDS}
\usage{
bb_load_cloud_counts(
  pipestance_path = NULL,
  specimen = NULL,
  umi_cutoff = 100,
  allowed_data_types = c("Gene Expression", "Antibody Capture"),
  sample_metadata_tbl = NULL
)
}
\arguments{
\item{pipestance_path}{A character string of the file path to the multi pipestance directory}

\item{specimen}{The biological specimen name.}

\item{umi_cutoff}{Don't import cells with fewer UMIs than this value}

\item{allowed_data_types}{Control type of data to import.}

\item{sample_metadata_tbl}{A tibble in wide format with one line.  Col names indicate metadata variables to add.}

\item{genome}{A character string for multi-genome samples; currently not working as originally intended so leave as null.}
}
\value{
A cell data set object.
}
\description{
Load 10X Cloud Pipestances Into CDS
}
