% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_load_multi_counts.R
\name{bb_load_multi_counts}
\alias{bb_load_multi_counts}
\title{Load 10X Multi data into a cds}
\usage{
bb_load_multi_counts(
  pipestance_path = NULL,
  specimen = NULL,
  genome = NULL,
  umi_cutoff = 100,
  allowed_data_types = c("Gene Expression", "Antibody Capture"),
  sample_metadata_nvps = NULL
)
}
\arguments{
\item{pipestance_path}{A character string of the file path to the multi pipestance directory}

\item{specimen}{The specimen name.}

\item{genome}{A character string for multi-genome samples}

\item{umi_cutoff}{Don't import cells with fewer UMIs than this value}

\item{allowed_data_types}{Control type of data to import.}

\item{sample_metadata_nvps}{A named vector or list of name+value pairs where name is the cell metadata column name to add and value is a sample-level metadata value.  These are applied to all cells in the sample (e.g.:  patient_id, treatment, timepoint).}
}
\value{
A cell data set object.
}
\description{
Load 10X Multi data into a cds
}
