% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_load_tenx.R
\name{bb_load_tenx_h5}
\alias{bb_load_tenx_h5}
\title{Load a 10X Genomics H5 File and Return a CDS}
\usage{
bb_load_tenx_h5(filename, sample_metadata_tbl = NULL, split_citeseq = FALSE)
}
\arguments{
\item{filename}{Path to the h5 file.}

\item{sample_metadata_tbl}{A single row data frame with sample metadata.  Usually this will be filtered from an experiment config file.}

\item{split_citeseq}{Option to retain citeseq data within the main experiment slot or split it out to an alternate slot.}
}
\value{
A cell data set.
}
\description{
This function reads a 10X Genomics H5 file and returns a cell_data_set or CDS.  The option to split citeseq data .
}
