% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ape_make.R
\name{bb_make_ape_transcript}
\alias{bb_make_ape_transcript}
\title{Make an Ape Transcriptome Object}
\usage{
bb_make_ape_transcript(query, transcriptome = c("hg38", "GRCz11"))
}
\arguments{
\item{query}{A specific ensembl transcript identifier.}

\item{transcriptome}{Genome/transcriptome reference to use, Default: c("hg38", "GRCz11")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
This function takes a specific ensembl transcript identifier, such as ENST00000348343.11, and gets the cDNA sequence from the corresponding transcriptomic reference.  This is returned as an Ape object with the UTR's and the CDS annotated as features.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[TxDb.Hsapiens.UCSC.hg38.knownGene]{TxDb.Hsapiens.UCSC.hg38.knownGene}}
\code{\link[BSgenome.Hsapiens.UCSC.hg38]{BSgenome.Hsapiens.UCSC.hg38}}
\code{\link[org.Hs.eg.db]{org.Hs.eg.db}}
\code{\link[blaseRdata]{character(0)}}
\code{\link[BSgenome.Drerio.UCSC.danRer11]{BSgenome.Drerio.UCSC.danRer11}}
\code{\link[org.Dr.eg.db]{org.Dr.eg.db}}
\code{\link[cli]{cli_abort}}
\code{\link[Biostrings]{matchPattern}}
}
