% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_funcs.R
\name{bb_makeTrace}
\alias{bb_makeTrace}
\title{Construct a Trace object from a Dataframe}
\usage{
bb_makeTrace(
  trace_data,
  startcol = NULL,
  endcol = NULL,
  genome = c("hg38", "danRer11"),
  seqname = NULL,
  plot_range = "trace",
  extend_left = 0,
  extend_right = 0,
  features = NULL,
  links = NULL
)
}
\arguments{
\item{trace_data}{A GRanges object or dataframe coercible to a GRanges object.  Should have metadata columns for "score" or "coverage" and "group".}

\item{startcol}{The dataframe column containing the start position.  Required only if trace_data is a dataframe.  Ignored if trace_data is a GRanges object.}

\item{endcol}{The dataframe column containing the end position.  Required only if trace_data is a dataframe.  Ignored if trace_data is a GRanges object.}

\item{genome}{The genome for the analysis.  Required.}

\item{seqname}{The sequence name, generally chromosome number.  Not required if trace_data is a GRanges object.}

\item{plot_range}{The bounding range for all slots in the object.  Everything will be trimmed to this range during generation of the object.  Can be shrunk after creation but not expanded.  There are three options for setting this:  1.  An explicit GRanges object can be provided.  2.  A valid gene name can be provided.  3.  Default:  the plot_range is set to match the initial range of the trace_data.  This last case will err if trace_data spans multiple chromosomes as is he case with unfiltered whole-sample bigwigs.}

\item{extend_left}{Bases to extend plot_range left, or upstream relative to the top strand.}

\item{extend_right}{Bases to extend plot_range right, or downstream relative to the top strand.}

\item{features}{A GRanges object with features to plot.  Should include a "type" metadata column for all feature types likely to be plotted.  Optional}

\item{links}{A GRanges object with cicero-style gene links.  Optional}
}
\description{
Use this function to create a trace object.  See details of the trace class for a complete description. Generally data can be in either NCBI or UCSC style.
}
