% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafeature_funcs.R
\name{bb_merge_narrowpeaks}
\alias{bb_merge_narrowpeaks}
\title{A Function To Merge Replicate Narrow Peaks GRanges}
\usage{
bb_merge_narrowpeaks(peaks_gr)
}
\description{
@description This function merges replicate peaks GRanges objects into 1.
@param peaks_gr A regular list of GRanges objects (Not a GRangesList).
@return A GRanges object
@export
@import IRanges GenomicRanges
}
