% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_monocle_anno.R
\name{bb_monocle_anno}
\alias{bb_monocle_anno}
\title{Annotate Single cell Data Using Monocle3}
\usage{
bb_monocle_anno(cds_qry, cds_ref, labels, suffix = "_ref")
}
\arguments{
\item{cds_qry}{A query cell data set.}

\item{cds_ref}{A reference cell data set.}

\item{labels}{A character vector of cell metadata column names to transfer.}

\item{suffix}{A character string of length 1 to append to all of the tranferred column names.  There is no checking for name conflicts, so use this sensibly to prevent overwriting preexisting columns.  Default = "_ref".}
}
\value{
A cell data set.
}
\description{
This function wraps the monocle3 method for data projection and label transfer into a function.  This function takes in reference and query cds objects plus a character vector of label column names and returns the query CDS with the new labels.
}
\details{
see https://cole-trapnell-lab.github.io/monocle3/docs/projection/
}
\seealso{
\code{\link[monocle3]{estimate_size_factors}}, \code{\link[monocle3]{preprocess_cds}}, \code{\link[monocle3]{reduce_dimension}}, \code{\link[monocle3]{save_transform_models}}, \code{\link[monocle3]{load_transform_models}}, \code{\link[monocle3]{preprocess_transform}}, \code{\link[monocle3]{reduce_dimension_transform}}, \code{\link[monocle3]{transfer_cell_labels}}, \code{\link[monocle3]{fix_missing_cell_labels}}
\code{\link[fs]{path}}
\code{\link[dplyr]{select}}, \code{\link[dplyr]{reexports}}
}
