% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_monocle_regression.R
\name{bb_monocle_regression}
\alias{bb_monocle_regression}
\title{A function to perform regression on single cell data.}
\usage{
bb_monocle_regression(
  cds,
  gene_or_genes,
  stratification_variable = NULL,
  stratification_value = NULL,
  form,
  linking_function = "negbinomial"
)
}
\arguments{
\item{cds}{A cell data set object.}

\item{gene_or_genes}{Genes to regress by.}

\item{stratification_variable}{Optional colData column to subset the cds by internal to the function.}

\item{stratification_value}{Optional value to stratify by.}

\item{form}{The regression formula in the form of "~var1+var2+..."}

\item{linking_function}{For the generalized linear model.}
}
\value{
A tibble containing the regression results.
}
\description{
A function to perform regression on single cell data.
}
