% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_monocle_regression.R
\name{bb_monocle_regression_better}
\alias{bb_monocle_regression_better}
\title{An Improved Function to Perform Regression on Single Cell Data.}
\usage{
bb_monocle_regression_better(
  cds,
  gene_or_genes,
  stratification_variable = NULL,
  stratification_value = NULL,
  form,
  linking_function = "negbinomial"
)
}
\arguments{
\item{cds}{A cell data set object.}

\item{gene_or_genes}{Genes to regress by.}

\item{stratification_variable}{Optional colData column to subset the cds by internal to the function.}

\item{stratification_value}{Optional value to stratify by.}

\item{form}{The regression formula in the form of "~var1+var2+..."}

\item{linking_function}{For the generalized linear model.}
}
\value{
A tibble containing the regression results.
}
\description{
This function replaces bb_monocle_regression.  It returns log2FoldChange instead of estimates.  The log2FoldChange value is the normalized_effect value returned from the monocle regression functions.
}
