% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_parseape.R
\name{bb_parseape}
\alias{bb_parseape}
\title{Parse a Genebank File and Construct an APE Object}
\usage{
bb_parseape(input_file)
}
\arguments{
\item{input_file}{The genebank/ape file to parse and construct into an instance of the Ape class.}
}
\value{
An Ape object
}
\description{
This is the main function for reading file in genebank/ape/equivalent format and generating an instance of the Ape class.  String manipulations are used to parse the input ape file.  Biostrings and GRanges functions are called to generate DNAStringSet and GRanges objects to store sequence and feature data, respectively.  The Ape constructor function is called internally at the end.
}
