% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_pseudotime.R
\name{bb_plot_genes_in_pseudotime}
\alias{bb_plot_genes_in_pseudotime}
\title{Plots expression for one or more genes as a function of pseudotime}
\usage{
bb_plot_genes_in_pseudotime(
  cds,
  gene_or_genes,
  pseudotime_dim,
  min_expr = NULL,
  cell_size = 0.75,
  nrow = NULL,
  ncol = 1,
  panel_order = NULL,
  color_cells_by = pseudotime_dim,
  trend_formula_df = 3,
  label_by_short_name = TRUE,
  vertical_jitter = NULL,
  horizontal_jitter = NULL,
  legend_title = NULL
)
}
\arguments{
\item{cds}{Cell data set to plot.}

\item{gene_or_genes}{Gene or genes for which to plot pseudotime.}

\item{pseudotime_dim}{The column holding the pseudotime dimension to plot along.}

\item{min_expr}{the minimum (untransformed) expression level to plot.}

\item{cell_size}{the size (in points) of each cell used in the plot.}

\item{nrow}{the number of rows used when laying out the panels for each
gene's expression.}

\item{ncol}{the number of columns used when laying out the panels for each
gene's expression}

\item{panel_order}{vector of gene names indicating the order in which genes
should be laid out (left-to-right, top-to-bottom). If
\code{label_by_short_name = TRUE}, use gene_short_name values, otherwise
use feature IDs.}

\item{color_cells_by}{the cell attribute (e.g. the column of colData(cds))
to be used to color each cell.  Defaults to the value provided for pseudotime_dim.}

\item{trend_formula_df}{degrees of freedom for the model formula used to fit the expression
trend over pseudotime.  The formulat takes the form of "~ splines::ns(pseudotime_dim, df = trend_formula_df)".}

\item{label_by_short_name}{label figure panels by gene_short_name (TRUE) or
feature ID (FALSE).}

\item{vertical_jitter}{A value passed to ggplot to jitter the points in the
vertical dimension. Prevents overplotting, and is particularly helpful for
rounded transcript count data.}

\item{horizontal_jitter}{A value passed to ggplot to jitter the points in
the horizontal dimension. Prevents overplotting, and is particularly
helpful for rounded transcript count data.}
}
\value{
a ggplot2 plot object
}
\description{
Plots expression for one or more genes as a function of pseudotime
}
