% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plotting.R
\name{bb_plot_heatmap_colDendro}
\alias{bb_plot_heatmap_colDendro}
\title{Plot a Heatmap Column Dendrogram}
\usage{
bb_plot_heatmap_colDendro(
  obj,
  side = c("top", "bottom", "left", "right"),
  linewidth = 0.5
)
}
\arguments{
\item{obj}{a Summarized Heatmap}

\item{side}{Orientation/side of the heatmap to put the dendrogram, Default: c("top", "bottom", "left", "right")}

\item{linewidth}{Weight of the dendrogram plot, Default: 0.5}
}
\value{
a ggplot
}
\description{
Takes in a SummarizedHeatmap object and returns a ggplot of the rowDendro slot.  This can be positioned with the side parameter.  Default is to position it on the left.  If flipping the heatmap so that the rows run vertically, you will need to change the side argument to top or bottom.
}
