% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plotting.R
\name{bb_plot_heatmap_colHighlight}
\alias{bb_plot_heatmap_colHighlight}
\title{Plot a Column Highlight}
\usage{
bb_plot_heatmap_colHighlight(
  obj,
  highlights = character(0),
  side = c("top", "bottom", "right", "left"),
  ...
)
}
\arguments{
\item{obj}{A summarized heatmap}

\item{highlights}{A vector of columns to highlight, Default: character(0)}

\item{side}{Side on which to put the higlight, Default: c("top", "bottom", "right", "left")}

\item{...}{Other arguments to pass to geom_text_repel}
}
\value{
a ggplot
}
\description{
Use geom_text_repel to selectively highlight some column names.  Useful when there are too many to highlight to be able to use the axis directly.
}
