% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plotting.R
\name{bb_plot_heatmap_main}
\alias{bb_plot_heatmap_main}
\title{Plot the Body of  Heatmap}
\usage{
bb_plot_heatmap_main(
  obj,
  tile_color = "white",
  high = "red3",
  mid = "white",
  low = "blue4",
  flip = FALSE
)
}
\arguments{
\item{obj}{A SummarizedHeatmap}

\item{tile_color}{Outline of the color tiles, Default: 'white'}

\item{high}{Color for high values, applied to scale_fill_gradient_2, Default: 'red3'}

\item{mid}{Color for mid values, applied to scale_fill_gradient_2, Default: 'white'}

\item{low}{Color for low values, applied to scale_fill_gradient_2, Default: 'blue4'}

\item{flip}{Whether to transpose the matrix, i.e. plot the rows as columns and columns as rows, Default: FALSE}
}
\value{
a ggplot
}
\description{
Takes in a Summarized Heatmap object and returns a ggplot of the matrix data.  Currently only supports clustering.
}
