% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plotting.R
\name{bb_plot_heatmap_rowData}
\alias{bb_plot_heatmap_rowData}
\title{Plot a SummarizedHeatmap rowData}
\usage{
bb_plot_heatmap_rowData(
  obj,
  tile_color = "white",
  vars = colnames(rowData(obj)),
  side = c("right", "left", "top", "bottom"),
  manual_pal = NULL
)
}
\arguments{
\item{obj}{A SummarizedHeatmap objectr}

\item{tile_color}{Color for the tile outlines, Default: 'white'}

\item{vars}{rowData variables to plot.  Supply a named vector to change the names shown on the axis and legend, Default: colnames(rowData(obj))}

\item{side}{Side on which to plotj, Default: c("right", "left", "top", "bottom")}

\item{manual_pal}{Color palette for filling the tiles, preferably a named vector, Default: NULL}
}
\value{
a ggplot
}
\description{
Use this function to create a plot annotating SummarizedHeatmap rowDAta
}
