% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plotting.R
\name{bb_plot_heatmap_rowHighlight}
\alias{bb_plot_heatmap_rowHighlight}
\title{Plot a Row Highlight}
\usage{
bb_plot_heatmap_rowHighlight(
  obj,
  highlights = character(0),
  side = c("top", "bottom", "right", "left"),
  ...
)
}
\arguments{
\item{obj}{A summarized Heatmap object}

\item{highlights}{A vector of rows to highlight, Default: character(0)}

\item{side}{Side on which to plot the highlight, Default: c("top", "bottom", "right", "left")}

\item{...}{Other arguments to pass to geom_text_repel}
}
\value{
a ggplot
}
\description{
Plots a selected row name from a SummarizedHeatmap object.  Useful when there are too many to highlight so you can't alter the plot axis.
}
