% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plot_rowData_col.R
\name{bb_plot_rowData_col}
\alias{bb_plot_rowData_col}
\title{A helper function to generate a data frame in the proper form for aggregate expression plotting with bb_gene_umap.}
\usage{
bb_plot_rowData_col(cds, rowData_col, filter_in = NULL, filter_out = NULL)
}
\arguments{
\item{cds}{CDS from which to extract the gene metadata.  Should be the same cds as the enclosing function.}

\item{rowData_col}{Gene metadata column to aggregate by.}

\item{filter_in}{Subset of values to focus on.  Each will become a facet in the final plot. Default is to keep everything except NA values.}

\item{filter_out}{Option to filter out any unwanted values. Default is to not filter out anything.}
}
\value{
A data frame in the format needed to pass into bb_gene_umap.
}
\description{
Use as argument for the "gene_or_genes" parameter for bb_gene_umap.
}
