% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_funcs.R
\name{bb_plot_trace_axis}
\alias{bb_plot_trace_axis}
\title{Plot The X Axis From A Trace Object}
\usage{
bb_plot_trace_axis(trace, xtitle = NULL)
}
\arguments{
\item{trace}{A Trace object.}

\item{xtitle}{An optional title for the X axis}
}
\description{
Generates the X axis for stacking other track plots on top of.
}
