% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_funcs.R
\name{bb_plot_trace_data}
\alias{bb_plot_trace_data}
\title{Plot The Trace Data From A Trace Object}
\usage{
bb_plot_trace_data(
  trace,
  yvar = "score",
  facet_var = NULL,
  colvar = NULL,
  pal = NULL
)
}
\arguments{
\item{trace}{A Trace object}

\item{yvar}{The variable that will become the Y axis.  Defaults to "score".}

\item{facet_var}{The variable describing data facets.  Each will be placed as a separate horizontal track with the value printed to the left. Optional but recommended.}

\item{colvar}{The variable to color traces by.  Optional but recommended.}

\item{pal}{A color palette.  Can also be added after the fact.}
}
\description{
A function to generate a line plot from tracklike genomic data.
}
