% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_funcs.R
\name{bb_plot_trace_feature}
\alias{bb_plot_trace_feature}
\title{Plot The Feature Data From A Trace Object}
\usage{
bb_plot_trace_feature(trace, type_to_plot)
}
\arguments{
\item{trace}{A Trace object.  Should have a metadata column in the features slot named "type".}

\item{type_to_plot}{Value of the type variable to plot.  This will be come the Y-axis label.}
}
\description{
A function to generate a feature plot from tracklike genomic data.
}
