% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_funcs.R
\name{bb_plot_trace_peaks}
\alias{bb_plot_trace_peaks}
\title{Plot The peak Data From A Trace Object}
\usage{
bb_plot_trace_peaks(
  trace,
  group_filter = NULL,
  group_variable = "group",
  pal = NULL
)
}
\arguments{
\item{trace}{A Trace object.}

\item{group_filter}{Optional value to filter the peak data by.  Should be a value from the "group" metadata variable in the trace object.}

\item{group_variable}{Optional metadatavariable to filter trace data by.  When imported from signac/seurat objects, this value defaults to "group", so that is the default here.  However if constructed manually, you may wish to apply filtering to another variable.  If so, apply it to this parameter.}

\item{fill_color}{The color to fill the peak graphics with.  Defaults to grey60.}
}
\description{
A function to generate a peak plot from tracklike genomic data.
}
