% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_plotfootprint.R
\name{bb_plotfootprint}
\alias{bb_plotfootprint}
\title{Plot motif footprinting results}
\usage{
bb_plotfootprint(
  object,
  features,
  alt_main_title = NULL,
  alt_color_title = NULL,
  legend_pos = "right",
  colorscale = NULL,
  assay = NULL,
  group.by = NULL,
  idents = NULL,
  label = TRUE,
  repel = TRUE,
  show.expected = TRUE,
  normalization = "subtract",
  label.top = 3,
  label.idents = NULL,
  fontsize = 14,
  linesize = 0.2
)
}
\arguments{
\item{object}{A Seurat object}

\item{features}{A vector of features to plot}

\item{alt_main_title}{Alternative title for the main plot.  Accepts markdown.}

\item{alt_color_title}{Alternative title for the color scale}

\item{legend_pos}{Position to place the legend}

\item{colorscale}{Named vector of colors to apply to the top plot.}

\item{assay}{Name of assay to use}

\item{group.by}{A grouping variable}

\item{idents}{Set of identities to include in the plot}

\item{label}{TRUE/FALSE value to control whether groups are labeled.}

\item{repel}{Repel labels from each other}

\item{show.expected}{Plot the expected Tn5 integration frequency below the
main footprint plot}

\item{normalization}{Method to normalize for Tn5 DNA sequence bias. Options
are "subtract", "divide", or NULL to perform no bias correction.}

\item{label.top}{Number of groups to label based on highest accessibility
in motif flanking region.}

\item{label.idents}{Vector of identities to label. If supplied,}

\item{fontsize}{Theme font size}

\item{linesize}{Size to draw the footprint lines
\code{label.top} will be ignored.}
}
\description{
Plot motif footprinting results
}
\concept{footprinting}
\concept{visualization}
