% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafeature_funcs.R
\name{bb_promoter_overlap}
\alias{bb_promoter_overlap}
\title{Calculate The Overlap Between Peaks and Promoters}
\usage{
bb_promoter_overlap(query, tss = c("hg38_tss", "dr11_tss"), width = 200)
}
\arguments{
\item{query}{A GRanges object containing peaks}

\item{tss}{The tss data base to use.  Must be one of "hg38_tss" or "dr11_tss"}

\item{width}{The width around the tss to evaluate.  Defaults to 200 bp.}
}
\value{
A list including overlap information and binomal test results.
}
\description{
For a given GRanges object containing peaks, determine how many peaks overlap promoters, how many promoters are overlapped by peaks and the significance of enrichment of query peaks relative to promoters.
}
