% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_pseudobulk.R
\name{bb_pseudobulk}
\alias{bb_pseudobulk}
\title{Run pseudobulk analysis using Deseq2}
\usage{
bb_pseudobulk(cds_deseq, replicate_variable, class_variable)
}
\arguments{
\item{cds_deseq}{The cell data set object subset to analyze}

\item{replicate_variable}{The colData column holding the biological replicate variable identifiers:  one unique identifier per biological replicate.}

\item{class_variable}{The colData column holding the class variable identifiers:  one unique identifier per class; two classes only}
}
\value{
A list of results from pseudobulk analysis
}
\description{
Run pseudobulk analysis using Deseq2
}
