% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_pseudotime.R
\name{bb_pseudotime}
\alias{bb_pseudotime}
\title{Learn Graph and Calculate Pseudotime}
\usage{
bb_pseudotime(
  cds,
  calculate_pseudotime = TRUE,
  cluster_variable,
  cluster_value,
  use_partition = TRUE,
  close_loop = TRUE,
  force_graph = FALSE
)
}
\arguments{
\item{cds}{The cell data set object to calculate pseudotime upon.  Does not yet accept seurat objects.}

\item{calculate_pseudotime}{Logical, whether to calculate the pseudotime dimension.  If false, will only run learn_graph, Default: TRUE}

\item{cluster_variable}{The cell metadata column from which the pseudotime = 0 cell will be selected.}

\item{cluster_value}{The value of cluster_variable that identifies a cluster.  The cell closest to the root node closest to the center of this cluster will have pseudotime of 0.}

\item{use_partition}{Logical; If TRUE, learn_graph will construct trajectories within partitions.  If FALSE, it will connect partitions, Default: TRUE}

\item{close_loop}{Logical; Whether learn_graph will close looping trajectories, Default: TRUE}

\item{force_graph}{Logical; If TRUE, the function will recalculate the graph., Default: FALSE}
}
\value{
A cell data set
}
\description{
This function determines a gene expression trajectory using \code{learn_graph} from monocle3 and then calculates pseudotime dimensions along this trajectory using \code{order_cells}.  So it is 2 functions wrapped into 1.  Usually we will not adjust the parameters for learn_graph with the possible exception of \code{close_loop} and \code{use_partition} which are also available in this function with the same defaults.  If you need to fine-tune the trajectory, use \code{monocle3::learn_graph} on the cds object first and then run this function to calculate pseudotime.  The graph learning will not be repeated on an object unless \code{force_graph} is set to \code{TRUE}.

If you just want to look at the trajectory graph and not calculate pseudotime, change \code{calculate_pseudotime} to FALSE, or run \code{monocle3::learn_graph}.

After the pseudotime values are calculated, they are handled differently than in monocle3.  In this function, they are copied from the hidden CDS slot and made an explicit cell metadata column.  Pseudotime needs a starting point or anchor.  There is no interactive option here as in monocle3.  To identify this starting point, you identify a cell metadata variable and provide it to \code{cluster_variable}.  This should identify a cohesive group of cells in UMAP space such as a leiden cluster, louvain cluster or partition.  Then provide a value corresponding to the cluster of interest to \code{cluster_value}.  The function will start pseudotime at the cell closest to the graph node in that cluster.  The pseudotime value column will be named automatically as a composite of the \code{cluster_variable} and \code{cluster_value} parameters.
}
\seealso{
\code{\link[cli]{cli_div}}, \code{\link[cli]{cli_alert}}
\code{\link[monocle3]{learn_graph}}, \code{\link[monocle3]{order_cells}}, \code{\link[monocle3]{pseudotime}}
\code{\link[SummarizedExperiment]{SummarizedExperiment-class}}
}
