% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_qc.R
\name{bb_qc}
\alias{bb_qc}
\title{A function to run qc tests on cds objects.}
\usage{
bb_qc(cds, cds_name, genome = c("human", "mouse", "zfish"))
}
\arguments{
\item{cds}{A cell data set object to run qc functions on}

\item{cds_name}{The name of the cds}

\item{genome}{The species to use for identifying mitochondrial genes.}
}
\value{
A list of qc objects
}
\description{
A function to run qc tests on cds objects.
}
