% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafeature_funcs.R
\name{bb_read_narrowpeak}
\alias{bb_read_narrowpeak}
\title{A Function to Read In Narrow Peaks Files}
\usage{
bb_read_narrowpeak(file)
}
\arguments{
\item{file}{The file path to the narrow peaks file.}
}
\value{
A GRanges object
}
\description{
This function reads the narrow peaks file (BED6 + 4 format) and turns it into a GRanges object.
}
