% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_rejoin.R
\name{bb_rejoin}
\alias{bb_rejoin}
\title{Rejoin qc and doubletfinder data to a cds object}
\usage{
bb_rejoin(cds, qc_data, doubletfinder_data)
}
\arguments{
\item{cds}{A cell data set object to rejoin to}

\item{qc_data}{A table of cell barcodes with qc data.  Can be extracted from bb_qc with purrr::map(qc_result, 1)}

\item{doubletfinder_data}{The doubletfinder result tbl}
}
\value{
A cell data set object with qc and doubletfinder data
}
\description{
Rejoin qc and doubletfinder data to a cds object
}
