% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_remove_dupes.R
\name{bb_remove_dupes}
\alias{bb_remove_dupes}
\title{Remove rows that have duplicates in a given column}
\usage{
bb_remove_dupes(data, column)
}
\arguments{
\item{data}{A tibble.}

\item{column}{A column to deduplicate}
}
\value{
A deduplicated tibble
}
\description{
Remove rows that have duplicates in a given column
}
