% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_renv_datapkg.R
\name{bb_renv_datapkg}
\alias{bb_renv_datapkg}
\title{Install Or Update A Local R Data Package}
\usage{
bb_renv_datapkg(path)
}
\arguments{
\item{path}{Path to a directory containing one or more versions of the same data package.  If a directory is specified, this function will compare the currently installed version to the latest availaible version in the directory.  If there is a newer version available (based on version number), it will install this version.  If a binary is specifically requested it will install that one. If the package hasn't been installed, it will install it.}
}
\value{
Returns nothing. Using renv, it installs the latest version of the binary datapackage or updates it if already installed.
}
\description{
Install Or Update A Local R Data Package
}
