% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_meta.R
\name{bb_rowmeta}
\alias{bb_rowmeta}
\title{Get Feature/Gene Metadata}
\usage{
bb_rowmeta(
  obj,
  row_name = "feature_id",
  experiment_type = "Gene Expression",
  assay = "RNA",
  cds = NULL
)
}
\arguments{
\item{obj}{A cell_data_set or Seurat object}

\item{row_name}{Optional name to provide for feature unique identifier, Default: 'feature_id'}

\item{experiment_type}{The experiment type to display.  Applies only to cds objects.  Commonly will be either "Gene Expression" or "Antibody Capture", Default: 'Gene Expression'}

\item{assay}{For a Seurat object, th feature assay to return.  CDS objects with alternative experiments are not supported, Default: 'RNA'}

\item{cds}{Provided for compatibility with prior versions, Default: NULL}
}
\value{
At tibble.
}
\description{
Take a cell_data_set or Seurat object and return the gene/feature metadata in the form of a tibble.  RNA is used as the default assay.
}
\details{
If a value is supplied for cds, a warning will be issued and the function will pass the value of cds to obj.
}
