% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_seurat_anno.R
\name{bb_seurat_anno}
\alias{bb_seurat_anno}
\title{A function to generate automated cell labelings with Seurat}
\usage{
bb_seurat_anno(cds, reference)
}
\arguments{
\item{cds}{A cell data set object}

\item{reference}{Seurat reference data.}
}
\value{
A modified cds with Seurat cell assignments.
}
\description{
A function to generate automated cell labelings with Seurat
}
