% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citeseq.R
\name{bb_split_citeseq}
\alias{bb_split_citeseq}
\title{Split Antibody Capture Data into Alt Experiment}
\usage{
bb_split_citeseq(cds)
}
\arguments{
\item{cds}{the cell data set to split}
}
\value{
a new CDS
}
\description{
If you have cite-seq data together with gene expression data, this function will move the cite seq data to a new separate experiment.  It will use Seurat to normalize these data using the CLR method and store them in a new assay.
}
\seealso{
\code{\link[SummarizedExperiment]{SummarizedExperiment-class}}
}
