% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_tbl_to_matrix.R
\name{bb_tbl_to_matrix}
\alias{bb_tbl_to_matrix}
\title{Convert a wide-form tibble a matrix}
\usage{
bb_tbl_to_matrix(data)
}
\arguments{
\item{data}{A wide form tibble to convert to a matrix.  The first column will become the rownames.}
}
\value{
A matrix
}
\description{
Convert a wide-form tibble a matrix
}
