% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_triplecluster.R
\name{bb_triplecluster}
\alias{bb_triplecluster}
\title{A function to generate clusters from scRNA-seq data}
\usage{
bb_triplecluster(cds, n_top_markers = 50, outfile, n_cores = 8)
}
\arguments{
\item{n_top_markers}{Number of top markers to identify per cell group}

\item{outfile}{Name of a csv file to hold the top marker results.}

\item{n_cores}{Number of processor cores to use}

\item{cds_string}{A cell data set object}
}
\value{
A modified cell data set object with leiden, louvain, and partition metadata columns.  Outputs csv files with top markers into specified directory, or current working directory if NULL.
}
\description{
Based on Monocle3's Partitions, Leiden, and Louvain clustering methods.  Implemented mostly with default values.
}
