% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_datahelpers.R
\name{data_median_se}
\alias{data_median_se}
\title{Generate median +/- se stat object for jitter ggplot}
\usage{
data_median_se(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Generate median +/- se stat object for jitter ggplot
}
