% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_datahelpers.R
\name{data_summary_mean_sd}
\alias{data_summary_mean_sd}
\title{Generate mean +/- sd stat object for jitter ggplot}
\usage{
data_summary_mean_sd(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Generate mean +/- sd stat object for jitter ggplot
}
