% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_datahelpers.R
\name{data_summary_mean_se}
\alias{data_summary_mean_se}
\title{Generate mean +/- se stat object for jitter ggplot}
\usage{
data_summary_mean_se(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Generate mean +/- se stat object for jitter ggplot
}
