% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_datahelpers.R
\name{data_summary_median_iqr}
\alias{data_summary_median_iqr}
\title{Generate median +/- iqr stat object for jitter ggplot}
\usage{
data_summary_median_iqr(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Generate median +/- iqr stat object for jitter ggplot
}
