% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_datahelpers.R
\name{data_summary_median_mad}
\alias{data_summary_median_mad}
\title{Generate median +/- mad stat object for jitter ggplot}
\usage{
data_summary_median_mad(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Generate median +/- mad stat object for jitter ggplot
}
