% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_grcz11_ape.R
\name{dnastring_to_origin}
\alias{dnastring_to_origin}
\title{Format a DNAString as a Character Vector For Inclusion in Genebank File}
\usage{
dnastring_to_origin(dna)
}
\arguments{
\item{dna}{A DNAString or a DNAStringSet of length 1.}
}
\value{
A character vector formatted for an Ape object ORIGIN slot.
}
\description{
This function takes a DNAString or DNAStringSet object and converts to a character string.  Spaces, newlines and line numbers are introduced to reproduce a 60-nucleotide/80-character-wide genebank ORIGIN section.
}
