% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ape_class.R
\name{FEATURES}
\alias{FEATURES}
\title{Get the Features Slot from an Ape Object}
\usage{
FEATURES(x)
}
\description{
This function gets the features slot from an Ape object.  Depending on console settings this may not be easily readable.  Better to simply show the whole Ape Object.  This may be useful for debugging.
}
