% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cds.R
\name{filter_cds}
\alias{filter_cds}
\title{Filter a CDS}
\usage{
filter_cds(cds, cells = "all", genes = "all")
}
\arguments{
\item{cds}{The CDS to filter.}

\item{cells}{Optional:  a tibble of cell metadata for the cells you wish to keep.  Use bb_cellmeta(). Default: 'all'}

\item{genes}{Optional:  a tibble of gene metadata for the genes you wish to keep.  Use bb_rowmeta().  Default: 'all'}
}
\value{
A filtered CDS
}
\description{
This function provides a pipe-friendly method to filter cds objects.
}
