% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-sig-table.R
\name{geom_sig_annotations}
\alias{geom_sig_annotations}
\title{Build significance annotation geoms}
\usage{
geom_sig_annotations(
  annotation_data,
  draw_brackets = TRUE,
  text_family = NULL,
  text_face = NULL,
  text_colour = "black",
  bracket_colour = "black",
  bracket_linewidth = 0.4,
  bracket_linetype = 1,
  bracket_lineend = "round",
  vjust = 0
)
}
\arguments{
\item{annotation_data}{Output from \code{\link[=prepare_sig_annotations]{prepare_sig_annotations()}}.}

\item{draw_brackets}{Logical; draw brackets if \code{TRUE}.}

\item{text_family, text_face, text_colour}{Text styling parameters.}

\item{bracket_colour, bracket_linewidth, bracket_linetype}{Bracket styling.}

\item{bracket_lineend}{Line ending for bracket segments. Defaults to \code{"round"}.}

\item{vjust}{Vertical justification for text.}
}
\value{
A list of ggplot layers.
}
\description{
Build significance annotation geoms
}
