% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_split_violin}
\alias{geom_split_violin}
\title{geom_split_violin}
\usage{
geom_split_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "identity",
  ...,
  draw_quantiles = NULL,
  trim = TRUE,
  scale = "area",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetitc mappings, Default: NULL}

\item{data}{Data to display, Default: NULL}

\item{stat}{Statistic to plot, Default: 'ydensity'}

\item{position}{Position, Default: 'identity'}

\item{...}{extra arguments}

\item{draw_quantiles}{Quantiles to draw, Default: NULL}

\item{trim}{Trim ends?, Default: TRUE}

\item{scale}{Analogous to violin scale, Default: 'area'}

\item{na.rm}{Default: FALSE}

\item{show.legend}{Default: NA}

\item{inherit.aes}{Default: TRUE}
}
\description{
create a split violin geom
}
