% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseape.R
\name{granges_to_features}
\alias{granges_to_features}
\title{Format a GRanges Object as a Character Vector For Inclusion in Genebank File}
\usage{
granges_to_features(gr)
}
\arguments{
\item{gr}{A GRanges object.}
}
\value{
A character vector.
}
\description{
This function takes a GRanges object and returns a character vector.  Only 4 metadata fields from the GRanges object will be included:  locus_tag, type, fwdcolor and revcolor.  Locus_tag must be unique.  This will be checked by the Ape constructor.  This function should mostly be used internally in the construction and FEATURE-setting of instances of the Ape class.
}
