% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\docType{class}
\name{Image_-class}
\alias{Image_-class}
\alias{Image_}
\alias{Image-class}
\alias{Image}
\alias{show,Image_-method}
\alias{file_path}
\alias{file_path<-}
\alias{species}
\alias{species<-}
\alias{stage}
\alias{stage<-}
\alias{genetics}
\alias{genetics<-}
\alias{treatment}
\alias{treatment<-}
\alias{microscope}
\alias{microscope<-}
\alias{mag}
\alias{mag<-}
\alias{filter}
\alias{filter<-}
\alias{use}
\alias{use<-}
\alias{note}
\alias{note<-}
\title{An S4 Class for Holding Image Metadata}
\usage{
\S4method{show}{Image_}(object)

file_path(x)

file_path(x) <- value

species(x)

species(x) <- value

stage(x)

stage(x) <- value

genetics(x)

genetics(x) <- value

treatment(x)

treatment(x) <- value

microscope(x)

microscope(x) <- value

mag(x)

mag(x) <- value

filter(x)

filter(x) <- value

use(x)

use(x) <- value

note(x)

note(x) <- value
}
\description{
This is an S4 Class for holding the relevant metadata we need for key images.  The idea is to generate this object for each of the images we will use in a grant, paper or other important document.  That way when we want to reuse these images we know where they are.

One can construct a single image object using the Image() constructor method.  When called as such, it will open a file chooser to identify the file from the network drive.  Then it will provide an interactive menu to add the needed metadata (see below).

The workflow is to use this to create a new Image object when you are using a new key image in a grant or other document.  Then you will add the image to the image catalog using ImageCatalog.add.

The Image constructor provides several validation checks, including that the image file must be accessible.

Each slot has it's own getter and setter methods which are identical to the name of the slot.

\strong{It is critical that your images are stored in a common network drive.  Ideally this is X/Labs/Blaser/staff/keyence imaging data}

\strong{Avoid Duplication!! Please keep your all of your raw imaging data in X/Labs/Blaser/staff and subdirectories.}
}
\section{Slots}{

\describe{
\item{\code{file_path}}{Path to file.  Should start with ~/network/X/Labs/Blaser...}

\item{\code{species}}{The species being imaged.}

\item{\code{stage}}{The stage of the sample.  Options include various times in hpf plus other for other timepoints.}

\item{\code{genetics}}{Any genetic modifications.}

\item{\code{treatment}}{Any treatments performed.}

\item{\code{microscope}}{The microscope used.}

\item{\code{mag}}{Magnification}

\item{\code{filter}}{The filter or camera setup used.}

\item{\code{use}}{The document the image is being used in.}

\item{\code{note}}{Any additional notes.}
}}

