% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\docType{class}
\name{ImageCatalog-class}
\alias{ImageCatalog-class}
\alias{ImageCatalog}
\alias{show,ImageCatalog-method}
\alias{[,ImageCatalog,ANY,ANY,ANY-method}
\alias{[[,ImageCatalog,ANY,ANY-method}
\alias{$,ImageCatalog-method}
\alias{ImageCatalog.as_tibble}
\alias{ImageCatalog.write}
\alias{ImageCatalog.add}
\alias{ImageCatalog.delete}
\title{An S4 Image Catalog}
\usage{
\S4method{show}{ImageCatalog}(object)

\S4method{[}{ImageCatalog,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[[}{ImageCatalog,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{$}{ImageCatalog}(x, name)

ImageCatalog.as_tibble(image_catalog)

ImageCatalog.write(image_catalog, out)

ImageCatalog.add(image_catalog, image)

ImageCatalog.delete(image_catalog, hash)
}
\description{
This object holds all of the individual Image objects.

Methods are provided for
\itemize{
\item viewing as a tibble:  ImageCatalog.as_tibble
\item writing to tsv format:  ImageCatalog.write
\item adding an image:  ImageCatalog.add
\item deleting an image:  ImageCatalog.delete
\item subsetting and extracting:  brackets and double brackets
}

An ImageCatalog object can be made from a list of Image objects using ImageCatalog(list = \if{html}{\out{<list of images>}}).

More commonly, you will generate the ImageCatalog from a tsv catalog file. To make an image catalog this way, run ImageCatalog(catalog_path = \if{html}{\out{<path>}}).
}
