% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix.utils.no_grr.R
\name{merge.Matrix}
\alias{merge.Matrix}
\alias{join.Matrix}
\title{Merges two Matrices or matrix-like objects}
\usage{
merge.Matrix(
  x,
  y,
  by.x,
  by.y,
  all.x = TRUE,
  all.y = TRUE,
  out.class = class(x)[1],
  fill.x = ifelse(is(x, "sparseMatrix"), FALSE, NA),
  fill.y = fill.x,
  ...
)

join.Matrix(
  x,
  y,
  by.x,
  by.y,
  all.x = TRUE,
  all.y = TRUE,
  out.class = class(x)[1],
  fill.x = ifelse(is(x, "sparseMatrix"), FALSE, NA),
  fill.y = fill.x,
  ...
)
}
\arguments{
\item{x, y}{\code{Matrix} or matrix-like object}

\item{by.x}{vector indicating the names to match from \code{Matrix} x}

\item{by.y}{vector indicating the names to match from \code{Matrix} y}

\item{all.x}{logical; if \code{TRUE}, then each value in \code{x} will be
included even if it has no matching values in \code{y}}

\item{all.y}{logical; if \code{TRUE}, then each value in \code{y} will be
included even if it has no matching values in \code{x}}

\item{out.class}{the class of the output object.  Defaults to the class of x.
Note that some output classes are not possible due to R coercion
capabilities, such as converting a character matrix to a Matrix.}

\item{fill.x, fill.y}{the value to put in merged columns where there is no match.
Defaults to 0/FALSE for sparse matrices in order to preserve sparsity, NA for
all other classes}

\item{...}{arguments to be passed to or from methods.  Currently ignored}
}
\description{
Implementation of \code{\link{merge}} for \code{\link{Matrix}}.  By explicitly
calling \code{merge.Matrix} it will also work for \code{matrix}, for
\code{data.frame}, and \code{vector} objects as a much faster alternative to
the built-in \code{merge}.
}
\details{
#' \code{all.x/all.y} correspond to the four types of database joins in the
following way:

\describe{ \item{left}{\code{all.x=TRUE}, \code{all.y=FALSE}}
\item{right}{\code{all.x=FALSE}, \code{all.y=TRUE}}
\item{inner}{\code{all.x=FALSE}, \code{all.y=FALSE}}
\item{full}{\code{all.x=TRUE}, \code{all.y=TRUE}} }

Note that \code{NA} values will match other \code{NA} values.
}
