% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_monocle_disk.R
\name{save_monocle_disk}
\alias{save_monocle_disk}
\title{Save a monocle3 cds object on disk}
\usage{
save_monocle_disk(cds_disk, data_directory, extdata_directory, alt_name = NULL)
}
\arguments{
\item{cds_disk}{The cds object to save to disk.}

\item{data_directory}{The package data directory to save to.}

\item{extdata_directory}{The package extdata directory to save to.}

\item{alt_name}{An alternative name to save the cds under.  Useful if the function is used programmatically.}
}
\value{
nothing
}
\description{
CDS objects can be large and because they are normally stored in memory this can lead to prolonged loading time and possibly system crashes.  Use this function to save a monocle cds object to disk.  The expected usage is to provide the function a cds object to save, and two directories.  These directories should be within a datapackage project, but this is not strictly required.  The extdata directory is where the cds data will be saved.  This \emph{must} be within inst (so, inst/extdata) for it to be installed and loaded with the data package using project_data.  The data directory is where normal .rda files are saved for typical R objects.  This will be a placeholder object with the same name.  The purpose is to permit documentation of the cds object and to mitigate namespace conflicts.  When the real object is loaded, the placeholder is overwritten.
}
\seealso{
\code{\link[fs]{file_access}}, \code{\link[fs]{path}}, \code{\link[fs]{path_file}}
\code{\link[cli]{cli_abort}}, \code{\link[cli]{cli_alert}}
\code{\link[monocle3]{convert_counts_matrix}}, \code{\link[monocle3]{save_monocle_objects}}
\code{\link[DescTools]{SaveAs}}
}
