% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ape_class.R
\name{setFeatures}
\alias{setFeatures}
\title{Set the FEATURES Slot of a GRanges Object}
\usage{
setFeatures(x, gr)
}
\arguments{
\item{x}{An ape object}

\item{gr}{A GRanges object.  This object will become the new FEATURES and granges slots for the Ape object.  So if you want to keep the old features, the new features need to be appended using c(old_gr, new_gr) as the value for the gr argument.}
}
\description{
Set the FEATURES Slot of a GRanges Object
}
