% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_funcs.R
\docType{class}
\name{Trace-class}
\alias{Trace-class}
\alias{Trace}
\title{An S4 class to Hold Genome Track Data}
\description{
An instance of this class is created by calling "bb_makeTrace".  All slots in this object are coerced to GRanges objects.  It is easiest to have everything start in UCSC-formated GRanges format but this is not necessary.  Validation checks will make sure data, features, links, and gene models all are bound by the same plot range on the same chromosome of the same genome.  Currently hg38 and danRer11 are the supported genomes.  Use this class to plot coverage tracks from ATAC or similar experiments.  Link plotting is available for cicero-style links.
}
\section{Slots}{

\describe{
\item{\code{trace_data}}{A Granges object or a dataframe coercible to a GRanges object.  bb_makeTrace should be able to handle NCBI or UCSC formatting styles.  Data will be rewritten as the latter.  Trace data should have a metadata column for "score" or "coverage" intended to be plotted as a y-variable.  Trace data must be normalized and combined into a single GRanges or dataframe before loading.  A group variable should be included to annotate color and and facets in the final plotting.  Whole sample bigwig files can be converted to GRanges objects using import.bw from rtracklayer.  All track data is trimmed during import but pre-trimming to the approximate range desired will significantly speed up processing.}

\item{\code{features}}{A GRanges object containing features to plot in a track-style format.  This object should include a "type" column to enable plotting each type in its own track.  All features of a given type are plotted on the same line.}

\item{\code{links}}{A GRanges object with Cicero-style links.  Must be GRanges but can be in either NCBI or UCSC format.}

\item{\code{gene_model}}{The gene model for plotting.  Will be automatically generated by bb_makeTrace.}

\item{\code{plot_range}}{The master GRange for the whole object.  Validity checks and/or constructors ensure all other ranges are contained within.}
}}

