% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_funcs.R
\name{Trace.setData}
\alias{Trace.setData}
\title{Set the Trace Data Slot of a GRanges Object}
\usage{
Trace.setData(trace, gr)
}
\arguments{
\item{trace}{A trace object}

\item{gr}{A GRanges object.  This object will become the new trace_data.  If the range is smaller, it will trim the other slots to match.  Usually this is used to change range metadata only.}
}
\description{
Set the Trace Data Slot of a GRanges Object
}
