% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_funcs.R
\name{Trace.setFeatures}
\alias{Trace.setFeatures}
\title{Set the Features Slot of a GRanges Object}
\usage{
Trace.setFeatures(trace, gr)
}
\arguments{
\item{trace}{An ape object}

\item{gr}{A GRanges object.  This object will become the new plot features.}
}
\description{
Set the Features Slot of a GRanges Object
}
