% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_funcs.R
\name{Trace.setLinks}
\alias{Trace.setLinks}
\title{Set the Links Slot of a GRanges Object}
\usage{
Trace.setLinks(trace, gr)
}
\arguments{
\item{trace}{An ape object}

\item{gr}{A GRanges object.  This object will become the new plot links.}
}
\description{
Set the Links Slot of a GRanges Object
}
