% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_funcs.R
\name{Trace.setpeaks}
\alias{Trace.setpeaks}
\title{Set the peaks Slot of a GRanges Object}
\usage{
Trace.setpeaks(trace, gr)
}
\arguments{
\item{trace}{An ape object}

\item{gr}{A GRanges object.  This object will become the new plot peaks.}
}
\description{
Set the peaks Slot of a GRanges Object
}
