% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_funcs.R
\name{Trace.setRange}
\alias{Trace.setRange}
\title{Set the Plot Range Slot of a GRanges Object}
\usage{
Trace.setRange(trace, gr)
}
\arguments{
\item{trace}{A trace object}

\item{gr}{A GRanges object.  This object will become the new plot range.}
}
\description{
Set the Plot Range Slot of a GRanges Object
}
