% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wordhash}
\alias{wordhash}
\title{English Word Hash Table For Image Blinding}
\format{
A data frame with 226857 rows and 2 variables:
\describe{
\item{\code{index}}{integer }
\item{\code{word}}{character }
}
}
\usage{
wordhash
}
\description{
Used in bb_blind_images to generate human-readable file hashes.
}
\details{
From package lexicon; see data-raw/wordhash.R
}
\keyword{datasets}
