% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_blind_images.R
\name{bb_blind_images}
\alias{bb_blind_images}
\title{Make a Copy of Image Files and Rename With File Hashes in Blinded Folder}
\usage{
bb_blind_images(source_dir, output_dir)
}
\arguments{
\item{source_dir}{The linux-style file path of the image files to be blinded.}

\item{output_dir}{The linux-style file path for the blinded images.  This will be created by the function.}
}
\value{
nothing
}
\description{
Will copy and rename  the files and generate two files:  "blinding_key.csv" with the original and blinded file names, and "scoresheet.csv" with just the blinded filenames.  Add columns as needed to scoresheet, for example, runx_count.  Then run bb_unblind to rejoin scoresheet to the key and generate an unblinded result file.
}
