% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_blind_images.R
\name{bb_blind_images}
\alias{bb_blind_images}
\title{Make a Copy of Image Files and Rename With File Hashes in Blinded Folder}
\usage{
bb_blind_images(analysis_file, file_column, output_dir)
}
\arguments{
\item{analysis_file}{The analysis file for the experiment.  It should contain 1 line for every biological sample and should have a filename for every file to be blinded.  It should be read from a csv using read_csv and in the form of a tibble.  If necessary it should be filtered upstream of this function to contain only the samples which you want to have blinded together.  Generally you would filter out unuseable images.  Also you would usually only wish to blind images from treatment groups from the same clutch.  So if you had 2 treatment groups replicated in 3 clutches, you would filter appropriately to generate 3 sets of blinded images across treatment groups.}

\item{file_column}{The column name in the analysis_file with the files to be blinded.}

\item{output_dir}{The linux-style file path for the directory that will hold the blinded images.  The directory will be created by the function.}
}
\value{
nothing
}
\description{
Will copy and rename  the files and generate two files:  "blinding_key.csv" with the original and blinded file names, and "scoresheet.csv" with just the blinded filenames.  Add columns as needed to scoresheet, for example, runx_count.  Then run bb_unblind to rejoin scoresheet to the key and generate an unblinded result file.
}
