% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_fix_file_path.R
\name{bb_fix_file_path}
\alias{bb_fix_file_path}
\title{Convert windows filepath to linux-compatible.}
\usage{
bb_fix_file_path(x)
}
\arguments{
\item{x}{A character string filepath copied from Windows with escaped backslashes.}
}
\value{
A linux-compatible filepath
}
\description{
This will only work if the backslashes have already been escaped to double backslashes.  For example, readr automatically does this when reading in windows file paths from csv files. If you are using as a standalone function, run this in the terminal:  scan(what = "character, n = 1), paste the unmodified filepath at the blank line, and copy the modified file path as the argument, x.
}
