% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_goenrichment.R
\name{bb_goenrichment}
\alias{bb_goenrichment}
\title{GO Term Enrichment}
\usage{
bb_goenrichment(
  query,
  reference,
  group_pval = 0.01,
  go_db = c("org.Hs.eg.db", "org.Dr.eg.db", "org.Mm.eg.db")
)
}
\arguments{
\item{query}{A vector of gene names}

\item{reference}{The background gene list.  Usually will be rowData(cds_main).}

\item{group_pval}{P value to determine enrichment.  Default: 0.01.}

\item{go_db}{PARAM_DESCRIPTION, Default: c("org.Hs.eg.db", "org.Dr.eg.db", "org.Mm.eg.db")}
}
\value{
A list of items including the enrichment results.
}
\description{
A function to find enriched go terms from a query list of gene names relative to a reference list of gene names.
}
