% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_gosummary.R
\name{bb_gosummary}
\alias{bb_gosummary}
\title{A function to reduce go terms by semantic similarity}
\usage{
bb_gosummary(
  x,
  reduce_threshold = 0.8,
  go_db = c("org.Hs.eg.db", "org.Dr.eg.db", "org.Mm.eg.db")
)
}
\arguments{
\item{x}{A list go term enrichment results produced by bb_goenrichment.}

\item{reduce_threshold}{The degree of term reduction. 0 to 1.  Higher is more reduction.}

\item{go_db}{The database to query.  Choose from c("org.Hs.eg.db", "org.Dr.eg.db", "org.Mm.eg.db", ...).}
}
\value{
A list of items for downstream plotting
}
\description{
A function to reduce go terms by semantic similarity
}
