% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_funcs.R
\name{bb_plot_trace_links}
\alias{bb_plot_trace_links}
\title{Plot The Link Data From A Trace Object}
\usage{
bb_plot_trace_links(trace, cutoff = 0)
}
\arguments{
\item{trace}{A Trace object containing a valid links slot.}

\item{cutoff}{Score cutoff for link plotting.  Defaults to 0.}
}
\description{
A function to generate a link plot from tracklike genomic data.  Links will automatically be trimmed to lie entirely within the plot range.  An additional, optionsl score cutoff can be provided.
}
