% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_funcs.R
\name{bb_plot_trace_peaks}
\alias{bb_plot_trace_peaks}
\title{Plot The peak Data From A Trace Object}
\usage{
bb_plot_trace_peaks(trace, fill_color = "grey60")
}
\arguments{
\item{trace}{A Trace object.}

\item{fill_color}{The color to fill the peak graphics with.  Defaults to grey60.}
}
\description{
A function to generate a peak plot from tracklike genomic data.
}
