% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_print_full_stats.R
\name{bb_print_full_stats}
\alias{bb_print_full_stats}
\title{Print out a stats report}
\usage{
bb_print_full_stats(data, class_var, class1, class2, value_var, out = NULL)
}
\arguments{
\item{data}{A Tibble in tidy data format}

\item{class_var}{Column containing the class variable}

\item{class1}{The first class to compare}

\item{class2}{The second class to compare}

\item{value_var}{The column containing the values}

\item{out}{The text file to output}
}
\value{
A text file
}
\description{
Print out a stats report
}
