% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_renv_datapkg.R
\name{bb_renv_datapkg}
\alias{bb_renv_datapkg}
\title{Install Or Update A Local R Data Package}
\usage{
bb_renv_datapkg(path)
}
\arguments{
\item{path}{Path to a directory containing one or more versions of the same data package.  There must be only 1 set of packages versions in the directory e.g. "my.package_1.0.tar.gz", "my.package_1.1.tar.gz", etc.}
}
\value{
Returns nothing. Using renv, it installs the latest version of the binary datapackage or updates it if already installed.
}
\description{
Install Or Update A Local R Data Package
}
