% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_meta.R
\name{bb_rowmeta}
\alias{bb_rowmeta}
\title{A Function to Return CDS Row Metadata in Tibble Form}
\usage{
bb_rowmeta(cds, row_name = "feature_id")
}
\arguments{
\item{cds}{A cell data set object}

\item{row_name}{Optional name to provide for cell unique identifier.  Defaults to "feature_id"}
}
\value{
A tibble
}
\description{
A Function to Return CDS Row Metadata in Tibble Form
}
