% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_tbl_to_coldata.R
\name{bb_tbl_to_coldata}
\alias{bb_tbl_to_coldata}
\title{A function to add tbl columns to cds colData}
\usage{
bb_tbl_to_coldata(cds, min_tbl, join_col = "barcode_sample")
}
\arguments{
\item{cds}{A cell data set object}

\item{min_tbl}{A tibble containing only the columns you want to add plus one column for joining.}

\item{join_col}{The column containing the join information for the cds colData. Defaults to "barcode_sample".}
}
\value{
A celldataset
}
\description{
A function to add tbl columns to cds colData
}
