% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_tbl_to_rowdata.R
\name{bb_tbl_to_rowdata}
\alias{bb_tbl_to_rowdata}
\title{A Function To Add Tibble Columns To CDS RowData}
\usage{
bb_tbl_to_rowdata(cds, min_tbl, join_col = "feature_id")
}
\arguments{
\item{cds}{A cell data set object}

\item{min_tbl}{A tibble containing only the columns you want to add plus one column for joining.}

\item{join_col}{The column in min_tbl containing the join information for the cds rowData. Defaults to "feature_id".}
}
\value{
A celldataset
}
\description{
A Function To Add Tibble Columns To CDS RowData
}
