% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_unblind.R
\name{bb_unblind_images}
\alias{bb_unblind_images}
\title{Rejoin Blinded Scores to Original File Names}
\usage{
bb_unblind_images(
  directory,
  keyfile = "blinding_key.csv",
  scorefile = "scoresheet.csv",
  analysis_file,
  file_column
)
}
\arguments{
\item{directory}{The linux-style filepath of the folder containing the scoresheet and blinded key.}

\item{keyfile}{Optional:  filename of the key file.  Defaults to "blinding_key.csv".}

\item{scorefile}{Optional:  filename of the score file.  Defaults to "scoresheet.csv".}

\item{analysis_file}{Complete file path to the the unblinded main analysis sheet.  The function will will left_join analysis_file and unblinded results.  In the process, it will necessarily convert windows file paths to linux-style file paths.  Samples not included in the blinding should return with NA values for the added columns.  New data columns being added on from scoresheet should be unique relative to analysis_file.}

\item{file_column}{The column in analysis_file containing file paths for the files that were blinded.}
}
\value{
nothing
}
\description{
Will rejoin scoresheet and blinded key to produce unblinded results.  If you change the names of either of those files, they have to be provided as arguments to the function.  Otherwise keyfile and scorefile are optional.
}
