% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_unblind.R
\name{bb_unblind_images}
\alias{bb_unblind_images}
\title{Rejoin Blinded Scores to Original File Names}
\usage{
bb_unblind_images(
  directory,
  keyfile = "blinding_key.csv",
  scorefile = "scoresheet.csv"
)
}
\arguments{
\item{directory}{The linux-style filepath of the folder containing the scoresheet and blinded key.}

\item{keyfile}{Optional:  filename of the key file.  Defaults to "blinding_key.csv".}

\item{scorefile}{Optional:  filename of the score file.  Defaults to "scoresheet.csv".}
}
\value{
nothing
}
\description{
Will rejoin scoresheet and blinded key to produce unblinded results.  If you change the names of either of those files, they have to be provided as arguments to the function.  Otherwise keyfile and scorefile are optional.
}
