% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\docType{class}
\name{ImageCatalog-class}
\alias{ImageCatalog-class}
\alias{ImageCatalog}
\title{An S4 Image Catalog}
\description{
This object holds all of the individual Image objects.

Methods are provided for
\itemize{
\item viewing as a tibble:  as_tibble
\item writing to .json format:  ImageCatalog.write
\item adding an image:  ImageCatalog.add
\item deleting an image:  ImageCatalog.delete
\item subsetting and extracting:  brackets and double brackets
}

An ImageCatalog object can be made from a list of Image objects using ImageCatalog(list = \if{html}{\out{<list of images>}}).

More commonly, you will generate the ImageCatalog from a .json file. To make an image catalog this way, run ImageCatalog(json_path = \if{html}{\out{<path>}}).
}
