% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_internal.R
\docType{package}
\name{SeuratWrappers-package}
\alias{blaseRtools}
\alias{blaseRtools-package}
\alias{SeuratWrappers-package}
\title{blaseRtools: R Tools for Blaser Lab Data Analysis}
\description{
This is a repository of R tools for Single Cell RNA seq and other lab data analysis functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/blaserlab/blaseRtools}
  \item Report bugs at \url{https://github.com/blaserlab/blaseRtools/issues}
}

}
\author{
\strong{Maintainer}: Brad Blaser \email{bradley.blaser@osumc.edu} (\href{https://orcid.org/0000-0002-3168-5423}{ORCID})

}
